rename project, create docker files and docker compose files for containerization
