+++ bytes_api/.dockerignore
@@ -1,0 +1,1 @@
+virtualenv
