+++ bytes_api/bytes/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class MealspirationConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'mealspiration'
+    name = 'bytes'
