+++ bytes_api/bytes/serializers.py
@@ -1,6 +1,7 @@
 from .models import CustomUser
 from rest_framework import serializers
 
+
 class CustomUserSerializer(serializers.ModelSerializer):
     class Meta:
         model = CustomUser
