+++ bytes_client/public/manifest.json
