+++ bytes_client/src/App.js
@@ -1,20 +1,22 @@
 import './App.css';
 import axios from 'axios';
 function App() {
-  axios.defaults.baseURL = 'http://127.0.0.1:8000'
+  axios.defaults.baseURL = 'http://127.0.0.1:8000/bytes_api/';
 
   const sendRequest = () => {
-    axios.get('mealspiration_api/users/', {
+    axios.get('users/', {
         format: 'json',
     })
     .then(function (response) {
+      console.log(response);
+    }).catch(function(response) {
       console.log(response);
     })
   }
 
   // sendUser
   const sendUser = (email, favorited_items, username, password) => {
-    axios.post('mealspiration_api/users/', {
+    axios.post('users/', {
       email: email,
       favorited_items: favorited_items,
       username: username,
@@ -23,10 +25,12 @@ function App() {
   }
 
   const getSpecific = (id) => {
-    axios.get(`mealspiration_api/users/${id}`, {
+    axios.get(`users/${id}`, {
       format: 'json'
     }).then(function(response) {
       console.log(response);
+    }).catch(function(response) {
+      console.log(response.error);
     })
   }
 
@@ -35,7 +39,7 @@ function App() {
       <h1> Hello</h1>
       <button onClick={sendRequest}>Click me to send a request</button>
       <button onClick={() => getSpecific(2)}>Get a specific item</button>
-      <button onClick={() => sendUser('kphan2@nd.edu', null, 'kdp', 'test123')}>Click me to send a user</button>
+      <button onClick={() => sendUser('jaynephan@mgmail.com', null, 'jphan', 'test123')}>Click me to send a user</button>
     </div>
   );
 }
