+++ docker-compose.yml
@@ -1,0 +1,26 @@
+version: '3'
+
+services:
+  backend:
+    build:
+      context: ./bytes_api
+    command: gunicorn api.wsgi --bind 0.0.0.0:8000
+    ports:
+      - "8000:8000"
+  frontend:
+    build:
+      context: ./bytes_client
+    volumes:
+      - react_build:/bytes_client/build
+  nginx:
+    image: nginx:latest
+    ports:
+      - 80:8080
+    volumes:
+      - ./nginx/nginx-setup.conf:/etc/nginx/conf.d/default.conf:ro
+      - react_build:/var/www/react/
+    depends_on:
+      - backend
+      - frontend
+volumes:
+  react_build:
