+++ nginx/nginx-setup.conf
@@ -1,0 +1,15 @@
+upstream api {
+    server backend:8000;
+}
+server {
+    listen 8080;
+
+    location / {
+        root /var/www/react;
+    }
+
+    location /bytes_api {
+        proxy_pass http://api;
+        proxy_set_header Host $http_host;
+    }
+}
