--- requirements.txt
@@ -1,8 +1,0 @@
-asgiref==3.7.2
-Django==4.2.5
-django-cors-headers==4.2.0
-djangorestframework==3.14.0
-mysqlclient==2.2.0
-python-dotenv==1.0.0
-pytz==2023.3.post1
-sqlparse==0.4.4
